feat: add notify componrnt

+++ src/components/TableMainRow/TableMainRow.tsx
@@ -1,6 +1,7 @@
 import { useAppDispatch } from '../../app/hooks';
 import { setUpdateNoteToForm, toggleShowNoteForm } from '../../reducers/formReducer';
 import { deleteNote, updateNote } from '../../reducers/noteReducer';
+import { removeNotify, setNotify } from '../../reducers/notifyReducer';
 import notesServices from '../../services/noteServices';
 import './TableMainRow.scss'
 import TableMainRowForm from './TableMainRowForm'
@@ -21,23 +22,26 @@ const TableMainRow = ({ note }: TableMainRowProps) => {
   const date = dates ? dates.join() : ''
   const lastDate = dates ? dates[dates.length - 1] : ''
   const dispatch = useAppDispatch()
-  const createAt = new Date(note.createAt).toLocaleString('en-US', { year: 'numeric',month: 'short',day: 'numeric',})
+  const createAt = new Date(note.createAt).toLocaleString('en-US', { year: 'numeric', month: 'short', day: 'numeric', })
 
   const handleDelete = async (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault()
     try {
-      await notesServices.deleteNote(note.id)
-      dispatch(deleteNote(note.id))
-      console.log('handleDelete',note.id);
+      const ask = confirm('Do you want to delete a note ?')
+      if (ask) {
+        await notesServices.deleteNote(note.id)
+        dispatch(deleteNote(note.id))
+        console.log('handleDelete', note.id);
+      }
     }
     catch (exception) {
-      //   dispatch(setNotifyMessage('Wrong credentials'))
-      //   setTimeout(() => {
-      //     dispatch(setNotifyMessage(null))
-      //   }, 5000)
-      // }
+      dispatch(setNotify(exception))
+      setTimeout(() => {
+        dispatch(removeNotify())
+      }, 5000)
     }
   };
+
   const handleUpdate = (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault()
     const noteInfo = {
@@ -47,42 +51,41 @@ const TableMainRow = ({ note }: TableMainRowProps) => {
       category: note.category,
       content: note.content,
       isArchive: note.isArchive
-     }
-     dispatch(toggleShowNoteForm())
-     dispatch(setUpdateNoteToForm(noteInfo))
+    }
+    dispatch(toggleShowNoteForm())
+    dispatch(setUpdateNoteToForm(noteInfo))
   };
   const handleArchive = async (event: React.MouseEvent<HTMLButtonElement>) => {
     event.preventDefault()
     try {
       const newNote = {
-        ...note , isArchive: !note.isArchive
+        ...note, isArchive: !note.isArchive
       }
       const updatedNote = await notesServices.updateNote(newNote)
       dispatch(updateNote(updatedNote))
-      console.log('handleArchive',note.id);
+      console.log('handleArchive', note.id);
     }
     catch (exception) {
-      //   dispatch(setNotifyMessage('Wrong credentials'))
-      //   setTimeout(() => {
-      //     dispatch(setNotifyMessage(null))
-      //   }, 5000)
-      // }
+      dispatch(setNotify(exception))
+      setTimeout(() => {
+        dispatch(removeNotify())
+      }, 5000)
     }
   };
 
   return (
-      <TableMainRowForm
-        handleDelete={handleDelete}
-        handleUpdate={handleUpdate}
-        handleArchive={handleArchive}
-        name={note.name}
-        createAt={createAt}
-        category={note.category}
-        content={note.content}
-        date={date}
-        id={note.id} 
-        lastDate={lastDate}
-        />
+    <TableMainRowForm
+      handleDelete={handleDelete}
+      handleUpdate={handleUpdate}
+      handleArchive={handleArchive}
+      name={note.name}
+      createAt={createAt}
+      category={note.category}
+      content={note.content}
+      date={date}
+      id={note.id}
+      lastDate={lastDate}
+    />
   )
 }
